refactor:  Restructure product routes and fix getProducts function
